+++ source/Internal/Domain/Authentication/Exception/UnavailableSourceOfRandomnessException.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\EshopCommunity\Internal\Domain\Authentication\Exception;
+
+use Exception;
+
+class UnavailableSourceOfRandomnessException extends Exception
+{
+}